+++ CotEditor/Sources/EditorTextView+CursorMovement.swift
@@ -519,7 +519,8 @@ extension EditorTextView {
     /// delete to the beginning of visual line (command+delete)
     override func deleteToBeginningOfLine(_ sender: Any?) {
         
-        guard self.hasMultipleInsertions else { return super.deleteToBeginningOfLine(sender) }
+        // -> Do not invoke super even with a single selection because the behavior of
+        //    `moveToBeginningOfLineAndModifySelection` is different from the default implementation.
         
         self.moveToBeginningOfLineAndModifySelection(sender)
         self.deleteBackward(sender)
@@ -531,7 +532,7 @@ extension EditorTextView {
         
         guard self.hasMultipleInsertions else { return super.deleteWordBackward(sender) }
         
-        self.moveWordForwardAndModifySelection(sender)
+        self.moveWordBackwardAndModifySelection(sender)
         self.deleteBackward(sender)
     }